django2/api_example/requirements.txt
Django==2.0.8
djangorestframework==3.8.2
